Fix MesaRepository id and queries

+++ ms-restaurante/src/main/java/cl/hilton/restaurante/repository/MesaRepository.java
@@ -1,15 +1,15 @@
 package cl.hilton.restaurante.repository;
 
-
-import cl.hilton.restaurante.model.Mesa;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
 import java.util.List;
 import java.util.Optional;
 
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import cl.hilton.restaurante.model.Mesa;
+
 @Repository
-public interface MesaRepository extends JpaRepository<Mesa, Integer> {
+public interface MesaRepository extends JpaRepository<Mesa, Long> {
 
     Optional<Mesa> findByNumeroMesa(String numeroMesa);
 
@@ -19,7 +19,7 @@ public interface MesaRepository extends JpaRepository<Mesa, Integer> {
 
     List<Mesa> findByDisponible(Boolean disponible);
 
-    List<Mesa> findByCapacidadGreaterThanEqual(Short capacidad);
+    List<Mesa> findByCapacidadGreaterThanEqual(Integer capacidad);
 
     List<Mesa> findByZonaAndDisponible(String zona, Boolean disponible);
 }
